hoocons/requirements.txt
-e git+https://github.com/django-nonrel/django.git@nonrel-1.6#egg=Django
-e git+https://github.com/django-nonrel/mongodb-engine.git@master#egg=django_mongodb_engine
-e git+https://github.com/django-nonrel/djangotoolbox.git@master#egg=djangotoolbox
-e git+https://github.com/django-nonrel/django@nonrel-1.5
-e git+https://github.com/MongoEngine/django-mongoengine.git@master#egg=django_mongoengine

fabric
django-extensions

django-appconf
django-contrib-comments

markdown
pygments
django-pagedown

# django rest framework
djangorestframework
django-filter

# for bcrypt passwords
# bcrypt
Pygments
requests

# Templates
django_compressor
djangorestframework
django-rest-framework-mongoengine
markdown
django-filter

blinker

# Documentation
Sphinx

# Testing
mock

# Security
djangorestframework-jwt

# Database
mongoengine